implement OrdMap::diff and OrdSet::diff
These methods are analogous to the Unix diff command in that they
produce a patch representing the minimal set of operations needed to
make one set or map equal to another.  The implementation is optimized
to skip shared B-tree nodes, meaning even very large sets can be
compared quickly when they mostly share the same structure.

The main intended application is keeping state synchronized between
remote peers, which can be acheived by serializing and transmitting
patches as new revisions of the state are published.