Make new and default implementation not depend on random state
This oddly fixes issue #148.

I tried to compose a set in a set. Which failed miserably, because the
HashSet would  not have  the same  hash, because  of the  random state
hasher implementation. This  was done to prevent hand  crafted key DOS
attacks.

To get things  working, you can build hash maps  with exactly the same
deterministic hasher. This becomes  error prone and inconvenient quite
quickly. Sprinkeling ::with_hasher() everywhere.

So this commit generalizes the ::new() and Default implementation over
the hasher (S) implementing Default and BuildHasher traits.

This  makes  it  possible  to  instantiate  the  RandomState  specific
implementations in the  root lib module, while  keeping the default(),
new() and unit() implementations general for each hasher.

So if I want a MyHashmap<K,V> with a SpecificHasher implementing
Default:

type MyHashMap<K,V> = im::hashmap::HashMap<K,V,SpecificHasher>;

Then all functions  still work for the different hasher.  I would then
also have to reimplement the macros. But that is fine.